refactor: :wastebasket: remove duplicated FormInput
